Extract reusable governance helpers from the e2e upgrade flow
Most of upgrade_flow.rs was either generic proposal-flow plumbing or
upgrade-specific PTB construction — neither of which needs to live in
the test crate. Pulls the non-test pieces into the hashi crate so both
the e2e harness and the CLI can share them.

New `hashi::cli::upgrade` module with:
  - build_upgrade_package — wraps `sui move build --dump-bytecode-as-base64`
    and parses the output into a Publish + digest.
  - build_upgrade_execution_transaction — constructs the
    `execute + publish + finalize` PTB the upgrade proposal requires.
  - build_execute_proposal_transaction — generic non-upgrade proposal
    execution; takes execute_package_id separately so post-upgrade
    disable_version still routes through the new package.
  - extract_proposal_id_from_response — parse ProposalCreatedEvent.
  - extract_new_package_id_from_response — find the new package in the
    upgrade tx effects.

In `hashi::cli::client`:
  - Replace the UpdateConfig-specific build_vote_update_config_transaction
    / build_execute_update_config_transaction with a generic standalone
    build_vote_transaction. HashiClient::build_vote_transaction now
    delegates to it, removing the duplicate PTB construction.

The e2e `upgrade_flow.rs` and the TestNetworksBuilder on-chain config
override flow now go through these helpers, which drops ~300 lines of
duplicated PTB boilerplate and inline event parsing.

No behaviour change; no new CLI commands are wired up yet. Follow-up
work on the cli-governance-improvements branch will add pre-flight
checks (version +1 enforcement, quorum readiness, etc.) and an
execute-upgrade CLI command that uses these helpers.
