+++ crates/e2e-tests/src/upgrade_flow.rs
@@ -10,8 +10,13 @@
 use anyhow::Result;
 use hashi::cli::client::CreateProposalParams;
 use hashi::cli::client::build_create_proposal_transaction;
+use hashi::cli::client::build_vote_transaction;
+use hashi::cli::upgrade::build_execute_proposal_transaction;
+use hashi::cli::upgrade::build_upgrade_execution_transaction;
+use hashi::cli::upgrade::build_upgrade_package;
+use hashi::cli::upgrade::extract_new_package_id_from_response;
+use hashi::cli::upgrade::extract_proposal_id_from_response;
 use hashi::config::HashiIds;
-use hashi::sui_tx_executor::SUI_CLOCK_OBJECT_ID;
 use hashi::sui_tx_executor::SuiTxExecutor;
 use std::path::Path;
 use std::path::PathBuf;
@@ -19,9 +24,6 @@ use sui_sdk_types::Address;
 use sui_sdk_types::Identifier;
 use sui_sdk_types::StructTag;
 use sui_sdk_types::TypeTag;
-use sui_transaction_builder::Function;
-use sui_transaction_builder::ObjectInput;
-use sui_transaction_builder::TransactionBuilder;
 
 use crate::TestNetworks;
 use crate::sui_network::sui_binary;
@@ -94,57 +96,6 @@ pub fn prepare_upgrade_package(test_dir: &Path, original_package_id: Address) ->
     Ok(dst)
 }
 
-/// Build the upgraded package and return the compiled modules + digest.
-pub fn build_upgrade_package(
-    package_path: &Path,
-    client_config: Option<&Path>,
-) -> Result<(sui_sdk_types::Publish, Vec<u8>)> {
-    let mut cmd = std::process::Command::new(sui_binary());
-    cmd.arg("move");
-
-    if let Some(config) = client_config {
-        cmd.arg("--client.config").arg(config);
-    }
-
-    cmd.arg("-p")
-        .arg(package_path)
-        .arg("build")
-        .arg("-e")
-        .arg("testnet")
-        .arg("--dump-bytecode-as-base64");
-
-    let output = cmd.output()?;
-    anyhow::ensure!(
-        output.status.success(),
-        "sui move build failed:\nstdout: {}\nstderr: {}",
-        output.stdout.escape_ascii(),
-        output.stderr.escape_ascii()
-    );
-
-    #[derive(serde::Deserialize)]
-    struct MoveBuildOutput {
-        modules: Vec<String>,
-        dependencies: Vec<Address>,
-        digest: Vec<u8>,
-    }
-
-    let build_output: MoveBuildOutput = serde_json::from_slice(&output.stdout)?;
-    let digest = build_output.digest.clone();
-    let modules = build_output
-        .modules
-        .into_iter()
-        .map(|b64| <base64ct::Base64 as base64ct::Encoding>::decode_vec(&b64))
-        .collect::<Result<Vec<_>, _>>()?;
-
-    Ok((
-        sui_sdk_types::Publish {
-            modules,
-            dependencies: build_output.dependencies,
-        },
-        digest,
-    ))
-}
-
 /// Run the full upgrade lifecycle: prepare → build → propose → vote → execute+publish+finalize.
 ///
 /// Returns the new package ID on success.
@@ -171,7 +122,7 @@ pub async fn execute_full_upgrade(networks: &mut TestNetworks) -> Result<Address
 
     // 2. Build the upgrade
     tracing::info!("building upgrade package from {}", upgrade_path.display());
-    let (compiled, digest) = build_upgrade_package(&upgrade_path, client_config)?;
+    let (compiled, digest) = build_upgrade_package(sui_binary(), &upgrade_path, client_config)?;
     tracing::info!("upgrade package built, digest: {digest:?}");
 
     // 3. Propose the upgrade
@@ -189,17 +140,7 @@ pub async fn execute_full_upgrade(networks: &mut TestNetworks) -> Result<Address
         "create Upgrade proposal failed"
     );
 
-    let proposal_id = response
-        .transaction()
-        .events()
-        .events()
-        .iter()
-        .find(|e| e.contents().name().contains("ProposalCreatedEvent"))
-        .ok_or_else(|| anyhow::anyhow!("ProposalCreatedEvent not found"))
-        .and_then(|e| {
-            let (id, _ts): (Address, u64) = bcs::from_bytes(e.contents().value())?;
-            Ok(id)
-        })?;
+    let proposal_id = extract_proposal_id_from_response(&response)?;
     tracing::info!("upgrade proposal {proposal_id} created");
 
     // 4. All other nodes vote (upgrade requires 100% quorum)
@@ -211,28 +152,8 @@ pub async fn execute_full_upgrade(networks: &mut TestNetworks) -> Result<Address
     )));
 
     for executor in &mut executors[1..] {
-        let mut builder = TransactionBuilder::new();
-        let hashi_arg = builder.object(
-            ObjectInput::new(hashi_ids.hashi_object_id)
-                .as_shared()
-                .with_mutable(true),
-        );
-        let proposal_id_arg = builder.pure(&proposal_id);
-        let clock_arg = builder.object(
-            ObjectInput::new(SUI_CLOCK_OBJECT_ID)
-                .as_shared()
-                .with_mutable(false),
-        );
-        builder.move_call(
-            Function::new(
-                hashi_ids.package_id,
-                Identifier::from_static("proposal"),
-                Identifier::from_static("vote"),
-            )
-            .with_type_args(vec![upgrade_type_tag.clone()]),
-            vec![hashi_arg, proposal_id_arg, clock_arg],
-        );
-        let vote_resp = executor.execute(builder).await?;
+        let vote_tx = build_vote_transaction(hashi_ids, proposal_id, upgrade_type_tag.clone());
+        let vote_resp = executor.execute(vote_tx).await?;
         anyhow::ensure!(
             vote_resp.transaction().effects().status().success(),
             "vote on Upgrade proposal failed"
@@ -242,69 +163,15 @@ pub async fn execute_full_upgrade(networks: &mut TestNetworks) -> Result<Address
 
     // 5. Execute upgrade + publish + finalize in one PTB
     tracing::info!("executing upgrade (execute + publish + finalize in one PTB)...");
-    let mut builder = TransactionBuilder::new();
-    let hashi_arg = builder.object(
-        ObjectInput::new(hashi_ids.hashi_object_id)
-            .as_shared()
-            .with_mutable(true),
-    );
-    let proposal_id_arg = builder.pure(&proposal_id);
-    let clock_arg = builder.object(
-        ObjectInput::new(SUI_CLOCK_OBJECT_ID)
-            .as_shared()
-            .with_mutable(false),
-    );
-
-    // Step A: upgrade::execute → UpgradeTicket
-    let ticket = builder.move_call(
-        Function::new(
-            hashi_ids.package_id,
-            Identifier::from_static("upgrade"),
-            Identifier::from_static("execute"),
-        ),
-        vec![hashi_arg, proposal_id_arg, clock_arg],
-    );
-
-    // Step B: publish upgrade → UpgradeReceipt
-    let receipt = builder.upgrade(
-        compiled.modules,
-        compiled.dependencies,
-        hashi_ids.package_id,
-        ticket,
-    );
-
-    // Step C: finalize_upgrade
-    let hashi_arg2 = builder.object(
-        ObjectInput::new(hashi_ids.hashi_object_id)
-            .as_shared()
-            .with_mutable(true),
-    );
-    builder.move_call(
-        Function::new(
-            hashi_ids.package_id,
-            Identifier::from_static("upgrade"),
-            Identifier::from_static("finalize_upgrade"),
-        ),
-        vec![hashi_arg2, receipt],
-    );
-
-    let upgrade_resp = executors[0].execute(builder).await?;
+    let upgrade_tx = build_upgrade_execution_transaction(hashi_ids, proposal_id, compiled);
+    let upgrade_resp = executors[0].execute(upgrade_tx).await?;
     anyhow::ensure!(
         upgrade_resp.transaction().effects().status().success(),
         "upgrade execute+publish+finalize failed: {:?}",
         upgrade_resp.transaction().effects().status()
     );
 
-    let new_package_id = upgrade_resp
-        .transaction()
-        .effects()
-        .changed_objects()
-        .iter()
-        .find(|o| o.object_type() == "package")
-        .ok_or_else(|| anyhow::anyhow!("new package not found in upgrade effects"))?
-        .object_id()
-        .parse::<Address>()?;
-
+    let new_package_id = extract_new_package_id_from_response(&upgrade_resp)?;
     tracing::info!("upgrade complete! new package: {new_package_id}");
     Ok(new_package_id)
 }
@@ -333,17 +200,7 @@ pub async fn disable_version(
         "create DisableVersion proposal failed"
     );
 
-    let proposal_id = response
-        .transaction()
-        .events()
-        .events()
-        .iter()
-        .find(|e| e.contents().name().contains("ProposalCreatedEvent"))
-        .ok_or_else(|| anyhow::anyhow!("ProposalCreatedEvent not found"))
-        .and_then(|e| {
-            let (id, _ts): (Address, u64) = bcs::from_bytes(e.contents().value())?;
-            Ok(id)
-        })?;
+    let proposal_id = extract_proposal_id_from_response(&response)?;
 
     let disable_version_type = TypeTag::Struct(Box::new(StructTag::new(
         hashi_ids.package_id,
@@ -353,55 +210,21 @@ pub async fn disable_version(
     )));
 
     for executor in &mut executors[1..] {
-        let mut builder = TransactionBuilder::new();
-        let hashi_arg = builder.object(
-            ObjectInput::new(hashi_ids.hashi_object_id)
-                .as_shared()
-                .with_mutable(true),
-        );
-        let proposal_id_arg = builder.pure(&proposal_id);
-        let clock_arg = builder.object(
-            ObjectInput::new(SUI_CLOCK_OBJECT_ID)
-                .as_shared()
-                .with_mutable(false),
-        );
-        builder.move_call(
-            Function::new(
-                hashi_ids.package_id,
-                Identifier::from_static("proposal"),
-                Identifier::from_static("vote"),
-            )
-            .with_type_args(vec![disable_version_type.clone()]),
-            vec![hashi_arg, proposal_id_arg, clock_arg],
-        );
-        let vote_resp = executor.execute(builder).await?;
+        let vote_tx = build_vote_transaction(hashi_ids, proposal_id, disable_version_type.clone());
+        let vote_resp = executor.execute(vote_tx).await?;
         anyhow::ensure!(
             vote_resp.transaction().effects().status().success(),
             "vote on DisableVersion proposal failed"
         );
     }
 
-    let mut builder = TransactionBuilder::new();
-    let hashi_arg = builder.object(
-        ObjectInput::new(hashi_ids.hashi_object_id)
-            .as_shared()
-            .with_mutable(true),
-    );
-    let proposal_id_arg = builder.pure(&proposal_id);
-    let clock_arg = builder.object(
-        ObjectInput::new(SUI_CLOCK_OBJECT_ID)
-            .as_shared()
-            .with_mutable(false),
-    );
-    builder.move_call(
-        Function::new(
-            execute_package_id,
-            Identifier::from_static("disable_version"),
-            Identifier::from_static("execute"),
-        ),
-        vec![hashi_arg, proposal_id_arg, clock_arg],
-    );
-    let exec_resp = executors[0].execute(builder).await?;
+    let execute_tx = build_execute_proposal_transaction(
+        hashi_ids,
+        proposal_id,
+        execute_package_id,
+        "disable_version",
+    )?;
+    let exec_resp = executors[0].execute(execute_tx).await?;
     anyhow::ensure!(
         exec_resp.transaction().effects().status().success(),
         "execute DisableVersion proposal failed"
